apps/ml-server/pyproject.toml
[project]
name = "ml-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.1",
    "joblib>=1.5.2",
    "onnxruntime>=1.22.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "uvicorn[standard]>=0.35.0",
    "xgboost>=3.0.5",
]
